Text outline support in TextField & TextEditor
Measured using memps on TM1 target by creating 500 TextFields.
Memory consumption is increased by 9% after outline is enabled.
Without outiline, the memory comsumption is almost the same as
before.

TextField with text only (no outline):

Before applying this patch:

  PID     CODE     DATA     PEAK     PSS       3D      GEM(PSS)
11272    13572    75120    88692    76920    58860     5568

After applying this patch:

  PID     CODE     DATA     PEAK     PSS       3D      GEM(PSS)
 5782    13580    75244    88824    77055    58508     5568

TextField with text and outline:

After applying this patch:

  PID     CODE     DATA     PEAK     PSS       3D      GEM(PSS)
 5851    13588    82056    95644    83874    63364     5568

+++ dali-toolkit/internal/text/text-controller.cpp
@@ -1182,14 +1182,14 @@ const Vector4& Controller::GetOutlineColor() const
   return mImpl->mModel->mVisualModel->GetOutlineColor();
 }
 
-void Controller::SetOutlineWidth( float width )
+void Controller::SetOutlineWidth( unsigned int width )
 {
   mImpl->mModel->mVisualModel->SetOutlineWidth( width );
 
   mImpl->RequestRelayout();
 }
 
-float Controller::GetOutlineWidth() const
+unsigned int Controller::GetOutlineWidth() const
 {
   return mImpl->mModel->mVisualModel->GetOutlineWidth();
 }
